refactor: split oversized feature components
Refs #29
