+++ components/features/chat/chat-utils.ts
@@ -1,0 +1,6 @@
+export function generateChatId(): string {
+  if (typeof crypto !== "undefined" && "randomUUID" in crypto) {
+    return crypto.randomUUID();
+  }
+  return Math.random().toString(36).slice(2) + Date.now().toString(36);
+}
