+++ components/features/sessions/sidebar-state.tsx
@@ -1,0 +1,41 @@
+"use client";
+
+import {
+  createContext,
+  useContext,
+  useMemo,
+  useState,
+  type ReactNode,
+} from "react";
+
+type SidebarContextValue = {
+  open: boolean;
+  setOpen: (open: boolean) => void;
+  toggle: () => void;
+};
+
+const SidebarContext = createContext<SidebarContextValue | null>(null);
+
+export function SidebarProvider({ children }: { children: ReactNode }) {
+  const [open, setOpen] = useState(true);
+  const value = useMemo(
+    () => ({
+      open,
+      setOpen,
+      toggle: () => setOpen((next) => !next),
+    }),
+    [open],
+  );
+
+  return (
+    <SidebarContext.Provider value={value}>{children}</SidebarContext.Provider>
+  );
+}
+
+export function useSidebar() {
+  const value = useContext(SidebarContext);
+  if (!value) {
+    throw new Error("useSidebar must be used within SidebarProvider");
+  }
+  return value;
+}
